AIO CPP 빌드: -> 'const' 를 추가했습니다 -> 'array_list' 를 만들었습니다

+++ lib4aio/lib4aio_cpp_sources/utils/memory_utils/memory_utils.cpp
@@ -1,9 +1,10 @@
+#include <malloc.h>
+#include <stdlib.h>
+
 namespace lib4aio
 {
 
-#include <stdlib.h>
-
-    void *new_object(const size_t size_of_object)
+    void *new_object(const unsigned size_of_object)
     {
         void *object = calloc(1, size_of_object);
         if (!object) {
@@ -13,7 +14,7 @@ namespace lib4aio
         return object;
     }
 
-    void *new_object_array(const size_t number_of_elements, const size_t size_of_object)
+    void *new_object_array(const unsigned number_of_elements, const unsigned size_of_object)
     {
         void *object_array = calloc(number_of_elements, size_of_object);
         if (!object_array) {
@@ -23,7 +24,11 @@ namespace lib4aio
         return object_array;
     }
 
-    void *reallocate_object_array(void *object_array, const size_t new_number_of_elements, const size_t size_of_object)
+    void *reallocate_object_array(
+            void *object_array,
+            const unsigned new_number_of_elements,
+            const unsigned size_of_object
+    )
     {
         object_array = realloc(object_array, new_number_of_elements * sizeof(size_of_object));
         if (!object_array) {
